+++ src/components/SarNFTPortfolio/Portfolio.tsx
@@ -1,5 +1,5 @@
 import { formatEther } from '@ethersproject/units';
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Position } from 'src/state/psarstake/hooks';
 import { Box } from '../Box';
@@ -15,6 +15,8 @@ interface Props {
 }
 
 export default function Portfolio({ itemsPerPage = 12, positions, onSelectPosition }: Props) {
+  const [pageCount, setPageCount] = useState(0);
+  const [itemOffset, setItemOffset] = useState(0);
   const [currentItems, setCurrentItems] = useState(positions.slice(0, itemsPerPage));
   const [selectedPositon, setSelectedPosition] = useState<Position | null>(null);
   const [selectedOption, setSelectedOption] = useState<string>('');
@@ -40,6 +42,17 @@ export default function Portfolio({ itemsPerPage = 12, positions, onSelectPositi
     } else {
       setCurrentItems(positions.slice(0, itemsPerPage));
     }
+  };
+
+  useEffect(() => {
+    const endOffset = itemOffset + itemsPerPage;
+    setCurrentItems(positions.slice(itemOffset, endOffset));
+    setPageCount(Math.ceil(positions.length / itemsPerPage));
+  }, [itemOffset, itemsPerPage]);
+
+  const handlePageClick = (event) => {
+    const newOffset = (event.selected * itemsPerPage) % positions.length;
+    setItemOffset(newOffset);
   };
 
   const renderItems = () => {
@@ -87,7 +100,7 @@ export default function Portfolio({ itemsPerPage = 12, positions, onSelectPositi
       </Box>
       <Box display="flex" flexDirection="column" flexGrow={1}>
         <Frame>{renderItems()}</Frame>
-        <Pagination items={currentItems} itemsPerPage={itemsPerPage} onChange={(items) => setCurrentItems(items)} />
+        <Pagination pageCount={pageCount} onPageChange={handlePageClick} />
       </Box>
     </Box>
   );
