+++ src/components/SarStakingWidget/AddStake/index.tsx
@@ -1,8 +1,11 @@
+import { formatEther } from '@ethersproject/units';
+import numeral from 'numeral';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Drawer } from 'src/components';
 import { Box } from 'src/components/Box';
 import { Button } from 'src/components/Button';
+import CurrencyLogo from 'src/components/CurrencyLogo';
+import { Stat } from 'src/components/Stat';
 import { Text } from 'src/components/Text';
 import { TextInput } from 'src/components/TextInput';
 import { ZERO_ADDRESS } from 'src/constants';
@@ -11,11 +14,12 @@ import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { ApprovalState } from 'src/hooks/useApproveCallback';
 import { useWalletModalToggle } from 'src/state/papplication/hooks';
 import { Position, useDerivativeSarStake, useSarStakeInfo } from 'src/state/psarstake/hooks';
-import { useTokenBalances } from 'src/state/pwallet/hooks';
+import { useTokenBalance } from 'src/state/pwallet/hooks';
+import ConfirmDrawer from '../ConfirmDrawer';
+import { Footer, Header, TokenRow } from '../ConfirmDrawer/styled';
 import Title from '../Title';
 import { Options } from '../types';
-import DrawerContent from './DrawerContent';
-import { Buttons, Root } from './styleds';
+import { Buttons, Root, Wrapper } from './styleds';
 
 interface Props {
   selectedOption: Options;
@@ -31,11 +35,10 @@ export default function AddStake({ selectedOption, selectedPosition, onChange }:
   const { account } = usePangolinWeb3();
 
   const png = PNG[chainId];
-  const tokensBalances = useTokenBalances(account ?? '', [png]);
-  const userPngBalance = tokensBalances[png.address];
+  const userPngBalance = useTokenBalance(account ?? '', png);
   const { t } = useTranslation();
 
-  const { APR } = useSarStakeInfo();
+  const { apr } = useSarStakeInfo();
 
   const toggleWalletModal = useWalletModalToggle();
 
@@ -54,6 +57,17 @@ export default function AddStake({ selectedOption, selectedPosition, onChange }:
     handleMax,
     onStake,
   } = useDerivativeSarStake(selectedPosition?.id);
+
+  const oldBalance = selectedPosition?.balance;
+  const newBalance = oldBalance?.add((parsedAmount?.raw ?? 0).toString()).add(selectedPosition?.pendingRewards ?? 0);
+
+  const newAPR = selectedPosition?.rewardRate
+    .mul(86400)
+    .mul(365)
+    .mul(100)
+    .div(newBalance ?? 1);
+
+  const weeklyPNG = selectedPosition?.rewardRate.mul(86400).mul(7);
 
   // check if user has gone through approval process, used to show two step buttons, reset on token change
   const [approvalSubmitted, setApprovalSubmitted] = useState<boolean>(false);
@@ -124,6 +138,45 @@ export default function AddStake({ selectedOption, selectedPosition, onChange }:
     }
   };
 
+  const ConfirmContent = (
+    <Wrapper paddingX="20px" paddingBottom="20px">
+      <Header>
+        <TokenRow>
+          <Text fontSize={24} fontWeight={500} color="text1" style={{ marginRight: '12px' }}>
+            {parsedAmount?.toSignificant(6) ?? 0}
+          </Text>
+          <CurrencyLogo currency={png} size={24} imageSize={48} />
+        </TokenRow>
+        <Box display="inline-grid" style={{ gridGap: '10px', gridTemplateColumns: 'auto auto' }}>
+          <Stat
+            title={t('sarStake.dollarValue')}
+            titlePosition="top"
+            stat={`${dollerWorth ?? 0}$`}
+            titleColor="text2"
+          />
+          <Stat title={t('sarStakeMore.newAPR')} titlePosition="top" stat={`${newAPR}%`} titleColor="text2" />
+        </Box>
+        <Box display="flex" flexDirection="row" justifyContent="space-between">
+          <Text color="text1">{t('sarStake.weeklyDistributed', { symbol: png.symbol })}</Text>
+          <Text color="text1">{numeral(formatEther(weeklyPNG ?? 0)).format('0.00a')}</Text>
+        </Box>
+        <Text color="text1" fontWeight={400} fontSize="14px" textAlign="center">
+          {t('sarStake.confirmDescription')}
+          <br />
+          <br />
+          {t('sarStakeMore.confirmDescription')}
+        </Text>
+      </Header>
+      <Footer>
+        <Box my={'10px'}>
+          <Button variant="primary" onClick={onStake}>
+            {t('sarStakeMore.add')}
+          </Button>
+        </Box>
+      </Footer>
+    </Wrapper>
+  );
+
   return (
     <Box>
       <Root>
@@ -159,7 +212,7 @@ export default function AddStake({ selectedOption, selectedPosition, onChange }:
             </Box>
             <Box>
               <Text color="text2">{t('sarStake.medianAPR')}</Text>
-              <Text color="text1">{`${(APR ?? '-').toString()}%`}</Text>
+              <Text color="text1">{`${(apr ?? '-').toString()}%`}</Text>
             </Box>
           </Box>
           <Text color="text1" fontWeight={400} fontSize="14px" textAlign="center">
@@ -172,25 +225,17 @@ export default function AddStake({ selectedOption, selectedPosition, onChange }:
         {renderButtons()}
       </Root>
 
-      <Drawer
+      <ConfirmDrawer
         title={stakeError || hash || attempting ? '' : t('sarStake.summary')}
-        isOpen={openDrawer}
+        isOpen={openDrawer && !!selectedPosition}
         onClose={handleConfirmDismiss}
-      >
-        {!!selectedPosition && (
-          <DrawerContent
-            stakeAmount={parsedAmount}
-            token={png}
-            dollerWorth={dollerWorth}
-            position={selectedPosition}
-            attemptingTxn={attempting}
-            txHash={hash}
-            onConfirm={onStake}
-            errorMessage={stakeError}
-            onClose={handleConfirmDismiss}
-          />
-        )}
-      </Drawer>
+        attemptingTxn={attempting}
+        txHash={hash}
+        errorMessage={stakeError}
+        pendingMessage={t('sarStakeMore.pending', { balance: parsedAmount?.toSignificant(2) ?? 0, symbol: png.symbol })}
+        successMessage={t('sarStake.successSubmit')}
+        confirmContent={ConfirmContent}
+      />
     </Box>
   );
 }
