+++ src/components/SarStakingWidget/Compound/index.tsx
@@ -1,24 +1,19 @@
 import { BigNumber } from '@ethersproject/bignumber';
 import { formatEther } from '@ethersproject/units';
 import numeral from 'numeral';
-import React, { useCallback, useContext, useEffect, useState } from 'react';
-import { AlertTriangle } from 'react-feather';
+import React, { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { ThemeContext } from 'styled-components';
-import CircleTick from 'src/assets/images/circleTick.svg';
-import { Drawer } from 'src/components';
 import { Box } from 'src/components/Box';
 import { Button } from 'src/components/Button';
-import { Loader } from 'src/components/Loader';
 import { Text } from 'src/components/Text';
 import { PNG } from 'src/constants/tokens';
 import { useChainId } from 'src/hooks';
 import { Position, useDerivativeSarCompound } from 'src/state/psarstake/hooks';
-import { getEtherscanLink } from 'src/utils';
 import { ToolTipText } from '../Claim/styleds';
+import ConfirmDrawer from '../ConfirmDrawer';
 import Title from '../Title';
 import { Options } from '../types';
-import { ErrorBox, ErrorWrapper, Link, Root, SubmittedWrapper } from './styleds';
+import { Root } from './styleds';
 interface Props {
   selectedOption: Options;
   selectedPosition: Position | null;
@@ -31,8 +26,6 @@ export default function Compound({ selectedOption, selectedPosition, onChange }:
   const chainId = useChainId();
 
   const { attempting, hash, compoundError, wrappedOnDismiss, onCompound } = useDerivativeSarCompound(selectedPosition);
-
-  const theme = useContext(ThemeContext);
 
   const { t } = useTranslation();
 
@@ -64,54 +57,9 @@ export default function Compound({ selectedOption, selectedPosition, onChange }:
     }
   }, [attempting]);
 
-  const PendingContent = <Loader size={100} label={t('sarCompound.pending')} />;
-
   const pendingRewards = selectedPosition?.pendingRewards ?? BigNumber.from('0');
 
   const png = PNG[chainId];
-
-  const ErroContent = (
-    <ErrorWrapper>
-      <ErrorBox>
-        <AlertTriangle color={theme.red1} style={{ strokeWidth: 1.5 }} size={64} />
-        <Text fontWeight={500} fontSize={16} color={'red1'} style={{ textAlign: 'center', width: '85%' }}>
-          {compoundError}
-        </Text>
-      </ErrorBox>
-      <Button variant="primary" onClick={handleConfirmDismiss}>
-        {t('transactionConfirmation.dismiss')}
-      </Button>
-    </ErrorWrapper>
-  );
-
-  const SubmittedContent = (
-    <SubmittedWrapper>
-      <Box display="flex" flexDirection="column" justifyContent="center" alignItems="center" paddingY={'20px'}>
-        <Box flex="1" display="flex" alignItems="center">
-          <img src={CircleTick} alt="circle-tick" />
-        </Box>
-        <Text fontSize={16} color="text1">
-          {t('sarCompound.successSubmit')}
-          <br />
-          {t('sarStake.yourAprRecalculated')}
-        </Text>
-        {chainId && hash && (
-          <Link
-            as="a"
-            fontWeight={500}
-            fontSize={14}
-            color={'primary'}
-            href={getEtherscanLink(chainId, hash, 'transaction')}
-          >
-            {t('transactionConfirmation.viewExplorer')}
-          </Link>
-        )}
-      </Box>
-      <Button variant="primary" onClick={handleConfirmDismiss}>
-        {t('transactionConfirmation.close')}
-      </Button>
-    </SubmittedWrapper>
-  );
 
   return (
     <Box>
@@ -156,9 +104,17 @@ export default function Compound({ selectedOption, selectedPosition, onChange }:
           {!selectedPosition ? t('sarStakeMore.choosePosition') : t('sarCompound.compound')}
         </Button>
       </Root>
-      <Drawer isOpen={openDrawer && !!selectedPosition} onClose={handleConfirmDismiss}>
-        {compoundError ? ErroContent : hash ? SubmittedContent : PendingContent}
-      </Drawer>
+
+      <ConfirmDrawer
+        isOpen={openDrawer && !!selectedPosition}
+        onClose={handleConfirmDismiss}
+        attemptingTxn={attempting}
+        txHash={hash}
+        errorMessage={compoundError}
+        pendingMessage={t('sarCompound.pending')}
+        successMessage={t('sarCompound.successSubmit')}
+        confirmContent={null}
+      />
     </Box>
   );
 }
